% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMAF.R
\name{filterMAF}
\alias{filterMAF}
\title{Filter MAF Data Frame}
\usage{
filterMAF(
  to_filter,
  tumor_only = FALSE,
  filter_column = c("PASS"),
  VAF_tumor = 0,
  VAF_control = 0,
  total_tumor_reads = 0,
  alt_tumor_reads = 0,
  cgi = FALSE,
  oncokb = FALSE,
  cgi_list = c("oncogenic (predicted)", "oncogenic (predicted and annotated)",
    "oncogenic (annotated)"),
  oncokb_list = c("Likely Oncogenic", "Oncogenic"),
  annott = c("HIGH", "MODERATE", "MODIFIER"),
  tumor_samples_out = NULL,
  control_samples_out = NULL
)
}
\arguments{
\item{to_filter}{A data frame containing the variants to be filtered in MAF format.}

\item{tumor_only}{A logical value indicating if the variant calling was performed in tumor-only mode. Default is FALSE, indicating it was performed in paired mode.}

\item{filter_column}{A character vector specifying the values to filter the `FILTER` column by. Default is "PASS".}

\item{VAF_tumor}{The minimum variant allele frequency in the tumor sample (from 0 to 1). Default is 0.}

\item{VAF_control}{The maximum variant allele frequency in the control sample (from 0 to 1). Default is 0.}

\item{total_tumor_reads}{The minimum number of reads in the tumor sample. Default is 0.}

\item{alt_tumor_reads}{The minimum number of alternative reads in the tumor sample. Default is 0.}

\item{cgi}{A logical value indicating if the CGI Summary column should be used to filter keeping labels included in cgi_list.}

\item{oncokb}{A logical value indicating if the OncoKB column should be used to filter keeping labels included in oncokb_list.}

\item{cgi_list}{Vector of CGI Summary labels to keep.}

\item{oncokb_list}{Vector of OncoKB labels to keep.}

\item{annott}{Vector of Annotation Impact labels from SnpEff to keep.}

\item{tumor_samples_out}{Vector of tumor samples names to remove.}

\item{control_samples_out}{Vector of control samples names to remove (this makes sense when the same tumor sample has been analysed using different control samples as germline control).}
}
\value{
A filtered data frame containing variants that meet the specified criteria.
}
\description{
This function filters a data frame of genetic variants based on specified criteria (VAF, reads & flags). It also allows the removal of tumor or control samples.
}
\examples{
filtered_df <- filterMAF(variants_df, tumor_only=TRUE, VAF_tumor=0.05, alt_tumor_reads=10)

}
