% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromParse2MAF.R
\name{fromParse2MAF}
\alias{fromParse2MAF}
\title{Read parse_variants Output and Convert To MAF}
\usage{
fromParse2MAF(path_to_parse, tumor_only = FALSE, oncokb = FALSE, cgi = FALSE)
}
\arguments{
\item{path_to_parse}{A vector of paths to the excel files to be read.}

\item{tumor_only}{A logical value indicating if variant calling was performed in tumor only mode. Default is FALSE, indicating it was performed in paired mode.}

\item{oncokb}{A logical value indicating whether OncoKB annotation was performed. Default is FALSE.}

\item{cgi}{A logical value indicating whether CGI annotation was performed. Default is FALSE.}
}
\value{
A data frame containing the processed variant data in MAF format.
}
\description{
This function reads variant data from one or more Excel files produced by parse_variants program and processes it in order to match MAF format specifications.
}
\examples{
# Example usage:
# variants_df <- fromParse2MAF(c("/path/to/file1","/path/to/file2"), tumor_only=TRUE)

}
