% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromParse2Onco.R
\name{fromParse2Onco}
\alias{fromParse2Onco}
\title{Convert output of parse_variants program into an oncoplot}
\usage{
fromParse2Onco(
  path_to_parse,
  tumor_only = FALSE,
  filter_column = c("PASS"),
  VAF_tumor = 0,
  VAF_control = 0,
  total_tumor_reads = 0,
  alt_tumor_reads = 0,
  cgi = FALSE,
  oncokb = FALSE,
  cgi_list = c("oncogenic (predicted)", "oncogenic (predicted and annotated)",
    "oncogenic (annotated)"),
  oncokb_list = c("Likely Oncogenic", "Oncogenic"),
  annott = c("HIGH", "MODERATE", "MODIFIER"),
  tumor_samples_out = NULL,
  control_samples_out = NULL,
  remove = TRUE,
  flags = FALSE,
  minimalMutations = 2,
  topgenes = 20000,
  nonSyn = c("Frame_Shift_Del", "Frame_Shift_Ins", "Splice_Site",
    "Translation_Start_Site", "Nonsense_Mutation", "Nonstop_Mutation", "In_Frame_Del",
    "In_Frame_Ins", "Missense_Mutation"),
  Missense_color = "#2a9134",
  Nonsense_color = "#ffca3a",
  Nonstop_color = "#000000",
  FrameDel_color = "blue",
  FrameIns_color = "purple",
  In_Frame_Ins_color = "lightblue",
  In_Frame_Del_color = "plum1",
  Tranlsation_Start_Site_color = "#ff0a54",
  Splice_site_color = "darkorange",
  Multihit_color = "#dab49d"
)
}
\arguments{
\item{path_to_parse}{A vector of paths to the excel files to be read.}

\item{tumor_only}{A logical value indicating if variant calling was performed in tumor only mode. Default is FALSE, indicating it was performed in paired mode.}

\item{filter_column}{A character vector specifying the values to filter the `FILTER` column by. Default is "PASS".}

\item{VAF_tumor}{The minimum variant allele frequency in the tumor sample (from 0 to 1). Default is 0.}

\item{VAF_control}{The maximum variant allele frequency in the control sample (from 0 to 1). Default is 0.}

\item{total_tumor_reads}{The minimum number of reads in the tumor sample. Default is 0.}

\item{alt_tumor_reads}{The minimum number of alternative reads in the tumor sample. Default is 0.}

\item{cgi}{A logical value indicating if the CGI Summary column should be used to filter keeping labels included in cgi_list.}

\item{oncokb}{A logical value indicating if the OncoKB column should be used to filter keeping labels included in oncokb_list.}

\item{cgi_list}{Vector of CGI Summary labels to keep.}

\item{oncokb_list}{Vector of OncoKB labels to keep.}

\item{annott}{Vector of Annotation Impact labels from SnpEff to keep.}

\item{tumor_samples_out}{Vector of tumor samples names to remove.}

\item{control_samples_out}{Vector of control samples names to remove (this makes sense when the same tumor sample has been analysed using different control samples as germline control).}

\item{remove}{A logical value indicating whether to remove duplicated variants. Equivalent to removeDuplicatedVariants in maftools::read.maf function. Default is TRUE.}

\item{flags}{A logical value indicating whether to remove FLAG genes. Equivalent to rmFlags in maftools::read.maf function. Default is FALSE.}

\item{minimalMutations}{Minimum number of samples mutated in a gene to be drawn. Equivalent to minMut in maftools::oncoplot function. Deafult is 2.}

\item{topgenes}{Number of top genes to be drawn. Equivalent to top in maftools::oncoplot function. Default = 20000.}

\item{nonSyn}{Vector of Variant Classifications to keep. Equivalent to vc_nonSyn in maftools::read.maf function. Default is c("Frame_Shift_Del", "Frame_Shift_Ins", "Splice_Site", "Translation_Start_Site","Nonsense_Mutation", "Nonstop_Mutation", "In_Frame_Del","In_Frame_Ins", "Missense_Mutation").}

\item{Missense_color}{Color for missense variants. Default is "#2a9134".}

\item{Nonsense_color}{Color for nonsense variants. Default is "#ffca3a".}

\item{Nonstop_color}{Color for nonstop variants. Default is "#000000".}

\item{FrameDel_color}{Color for frame shift deletions. Default is "blue".}

\item{FrameIns_color}{Color for frame shift insertions. Default is "purple".}

\item{In_Frame_Ins_color}{Color for in-frame insertions. Default is "lightblue".}

\item{In_Frame_Del_color}{Color for in-frame deletions. Default is "plum1".}

\item{Tranlsation_Start_Site_color}{Color for translation start sites. Default is "#ff0a54".}

\item{Splice_site_color}{Color for splicing sites. Default is "darkorange".}

\item{Multihit_color}{Color for multi-hit genes. Default is "#dab49d".}
}
\value{
An oncoplot object which is also saved as a png file.
}
\description{
This function reads one or more Excel files produced by parse_variants program and makes and oncoplot.
}
\examples{
oncobuddy <-fromParse2Onco(path_to_parse = c("/path/to/file1","/path/to/file2"), VAF_tumor=0.05, minimalMutations=3, Missense_color="#FF5733")

}
