% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imporandfilterbysamples.R
\name{importandfilterbysamples}
\alias{importandfilterbysamples}
\title{Parse and Combine Variant Data}
\usage{
importandfilterbysamples(minimum_samples_mutated)
}
\arguments{
\item{minimum_samples_mutated}{Minimum number of samples that must be mutated to be included in the filtered matrix.}
}
\value{
Matrix filtered by number of samples mutated
}
\description{
Import and Filter by Samples
}
\examples{
\dontrun{
filtered_matrix <- importandfilterbysamples(minimum_samples_mutated = 5)
}
}
