% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOncoplot.R
\name{makeOncoplot}
\alias{makeOncoplot}
\title{Make Oncoplot}
\usage{
makeOncoplot(
  Missense_color = "#2a9134",
  Nonsense_color = "#ffca3a",
  Nonstop_color = "#000000",
  FrameDel_color = "blue",
  FrameIns_color = "purple",
  In_Frame_Ins_color = "lightblue",
  In_Frame_Del_color = "plum1",
  Translation_Start_Site_color = "#ff0a54",
  Splice_site_color = "darkorange",
  Multihit_color = "#dab49d"
)
}
\arguments{
\item{Missense_color}{Color for missense variants. Default is "#2a9134".}

\item{Nonsense_color}{Color for nonsense variants. Default is "#ffca3a".}

\item{Nonstop_color}{Color for nonstop variants. Default is "#000000".}

\item{FrameDel_color}{Color for frame shift deletions. Default is "blue".}

\item{FrameIns_color}{Color for frame shift insertions. Default is "purple".}

\item{In_Frame_Ins_color}{Color for in-frame insertions. Default is "lightblue".}

\item{In_Frame_Del_color}{Color for in-frame deletions. Default is "plum1".}

\item{Translation_Start_Site_color}{Color for translation start sites. Default is "#ff0a54".}

\item{Splice_site_color}{Color for splicing sites. Default is "darkorange".}

\item{Multihit_color}{Color for multi-hit genes. Default is "#dab49d".}
}
\value{
An oncoplot object which is also saved as a png file.
}
\description{
This function imports an onco_matrix.txt file in the working directory and makes an oncoplot using the oncoPrint function in ComplexHeatmap package.
}
\examples{
oncoplot <- makeOncoplot(Missense_color="#FF5733", Nonsense_color="#33FF57")

}
