% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modclassvar.R
\name{modclassvartmb}
\alias{modclassvartmb}
\title{Modify and Classify Variants, Calculate TMB}
\usage{
modclassvartmb(
  to_class,
  remove = FALSE,
  minimum_samples_mutated = 1,
  flags = FALSE,
  minimalMutations = 2,
  topgenes = 20000,
  nonSyn = c("Frame_Shift_Del", "Frame_Shift_Ins", "Splice_Site",
    "Translation_Start_Site", "Nonsense_Mutation", "Nonstop_Mutation", "In_Frame_Del",
    "In_Frame_Ins", "Missense_Mutation")
)
}
\arguments{
\item{to_class}{A data frame containing the variants to be classified.}

\item{remove}{A logical value indicating whether to remove duplicated variants. Default is FALSE.}

\item{minimum_samples_mutated}{An integer specifying the minimum number of samples mutated to filter the matrix. Default is 1.}
}
\value{
A filtered matrix of the variant classifications.
}
\description{
This function modifies the classification of variants and generates summary plots.
}
\examples{
# Example usage:
# filtered_matrix <- modclassvar(variant_data, remove=TRUE, minimum_samples_mutated=5)

}
