% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareForOncoplot.R
\name{prepareForOncoplot}
\alias{prepareForOncoplot}
\title{Create oncomatrix from MAF Data Frame}
\usage{
prepareForOncoplot(
  maf_df,
  remove = TRUE,
  flags = FALSE,
  minimalMutations = 2,
  topgenes = 20000,
  nonSyn = c("Frame_Shift_Del", "Frame_Shift_Ins", "Splice_Site",
    "Translation_Start_Site", "Nonsense_Mutation", "Nonstop_Mutation", "In_Frame_Del",
    "In_Frame_Ins", "Missense_Mutation")
)
}
\arguments{
\item{maf_df}{A data frame in MAF format.}

\item{remove}{A logical value indicating whether to remove duplicated variants. Equivalent to removeDuplicatedVariants in maftools::read.maf function. Default is TRUE.}

\item{flags}{A logical value indicating whether to remove FLAG genes. Equivalent to rmFlags in maftools::read.maf function. Default is FALSE.}

\item{minimalMutations}{Minimum number of samples mutated in a gene to be drawn. Equivalent to minMut in maftools::oncoplot function. Default is 2.}

\item{topgenes}{Number of top genes to be drawn. Equivalent to top in maftools::oncoplot function. Default = 20000.}

\item{nonSyn}{Vector of Variant Classifications to keep. Equivalent to vc_nonSyn in maftools::read.maf function. Default is c("Frame_Shift_Del", "Frame_Shift_Ins", "Splice_Site", "Translation_Start_Site","Nonsense_Mutation", "Nonstop_Mutation", "In_Frame_Del","In_Frame_Ins", "Missense_Mutation").}
}
\value{
It prints summary plots and saves the oncomatrix (onco_matrix.txt) and the TMB table (TMB_table.txt) in the working directory.
}
\description{
This function takes a MAF dataframe, filters it and creates an oncomatrix. It also prints summary plots and creates a TMB table. It does so by executing functions in maftools package.
}
\examples{
# Example usage:
# prepareForOncoplot(filtered_df, minimalMutations=3)

}
