% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteringtable.R
\name{filtering_table}
\alias{filtering_table}
\title{Filtering Variants Data Frame}
\usage{
filtering_table(
  to_filter,
  tumor_only = FALSE,
  filter_column = c("PASS"),
  VAF_tumor = 0,
  VAF_control = 0,
  total_reads = 0,
  tumor_reads = 0,
  cgi_path = FALSE,
  oncokb_path = FALSE,
  cgi_list = c("oncogenic (predicted)", "oncogenic (predicted and annotated)"),
  oncokb_list = c("Likely Oncogenic", "Oncogenic"),
  annott = c("HIGH", "MODERATE", "MODIFIER")
)
}
\arguments{
\item{to_filter}{A data frame containing the variants to be filtered.}

\item{tumor_only}{A logical value indicating if only tumor samples should be considered. Default is FALSE.}

\item{filter_column}{A character vector specifying the values to filter the `FILTER` column by. Default is "PASS".}

\item{VAF_tumor}{A numeric threshold for the variant allele frequency (VAF) in the tumor sample. Default is 0.}

\item{VAF_control}{A numeric threshold for the variant allele frequency (VAF) in the control sample. Default is 0.}

\item{total_reads}{A numeric threshold for the total reads. Default is 0.}

\item{tumor_reads}{A numeric threshold for the alternative reads in the tumor sample. Default is 0.}

\item{cgi_path}{A boolean that defines if CGI column exists}

\item{oncokb_path}{A boolean that defines if oncokb column exists}
}
\value{
A filtered data frame containing variants that meet the specified criteria.
}
\description{
This function filters a data frame of genetic variants based on specified criteria (VAF, READS & FLAGS)
}
\examples{
# Example usage:
# filtered_df <- filtering_table(variants_df, tumor_only=TRUE, VAF_tumor=0.05, tumor_reads=10)
}
