% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FromZero2Plot.R
\name{fromparse2onco}
\alias{fromparse2onco}
\title{Convert Parsed Data to Oncoplot}
\usage{
fromparse2onco(
  path_to_parse,
  more_than_one,
  tumor_only,
  pattern_excel = FALSE,
  filter_column = c("PASS"),
  VAF_tumor = 0,
  VAF_control = 0,
  total_reads = 0,
  tumor_reads = 0,
  remove = FALSE,
  minimum_samples_mutated = 1,
  Missense_color = "#2a9134",
  Nonsense_color = "#ffca3a",
  Nonstop_color = "#000000",
  FrameDel_color = "blue",
  FrameIns_color = "purple",
  In_Frame_Ins_color = "lightblue",
  In_Frame_Del_color = "plum1",
  Tranlsation_Start_Site_color = "#ff0a54",
  Splice_site_color = "darkorange",
  Multihit_color = "#dab49d",
  flags = FALSE,
  nonSyn = c("Frame_Shift_Del", "Frame_Shift_Ins", "Splice_Site",
    "Translation_Start_Site", "Nonsense_Mutation", "Nonstop_Mutation", "In_Frame_Del",
    "In_Frame_Ins", "Missense_Mutation"),
  cgi_path = FALSE,
  oncokb_path = FALSE,
  cgi_list = c("oncogenic (predicted)", "oncogenic (predicted and annotated)"),
  oncokb_list = c("Likely Oncogenic", "Oncogenic"),
  annott = c("HIGH", "MODERATE", "MODIFIER"),
  oncokb = FALSE,
  cgi = FALSE,
  minimalMutations = 10,
  topgenes = 50
)
}
\arguments{
\item{path_to_parse}{Path to the parsed data file(s).}

\item{more_than_one}{Logical indicating whether there are multiple files to process.}

\item{tumor_only}{Logical indicating whether only tumor data should be considered.}

\item{pattern_excel}{Pattern for matching Excel files.}

\item{filter_column}{Filter column for variant filtering.}

\item{VAF_tumor}{Minimum VAF for tumor samples.}

\item{VAF_control}{Minimum VAF for control samples.}

\item{total_reads}{Minimum total reads.}

\item{tumor_reads}{Minimum tumor reads.}

\item{remove}{Logical indicating whether duplicated variants should be removed.}

\item{minimum_samples_mutated}{Minimum number of samples mutated for inclusion in the oncoplot.}

\item{Missense_color}{Color for missense variants. Default is "#2a9134".}

\item{Nonsense_color}{Color for nonsense variants. Default is "#ffca3a".}

\item{Nonstop_color}{Color for nonstop variants. Default is "#000000".}

\item{FrameDel_color}{Color for frame shift deletions. Default is "blue".}

\item{FrameIns_color}{Color for frame shift insertions. Default is "purple".}

\item{In_Frame_Ins_color}{Color for in-frame insertions. Default is "lightblue".}

\item{In_Frame_Del_color}{Color for in-frame deletions. Default is "plum1".}

\item{Tranlsation_Start_Site_color}{Color for translation start sites. Default is "#ff0a54".}

\item{Splice_site_color}{Color for splicing sites. Default is "darkorange".}

\item{Multihit_color}{Color for multi-hit genes. Default is "#dab49d".}

\item{minimalMutations}{Minimum number of mutations per gene to not be discarded. Default is 10.}

\item{topgenes}{Number of genes to show in the oncoplot}
}
\value{
An object to represent an oncoplot or an oncoplot
}
\description{
Convert Parsed Data to Oncoplot
}
\examples{
oncobuddy <-fromparse2onco("/path/to/parse/",TRUE,TRUE,pattern_excel="pattern")
}
